Correctly parsing negative numbers through hand modification of lexer.js
Note: the definitions for constants are defined in grammar.jison,
however in the gulp step it uses the pre-built contents of lexer.js. How
lexer.js is created is not fully understood, except that it was done
with the npm package jison-lex, which is why this commit adds it to the
package.json file

I manually added the correct regexes to the line in lexer.js. Ideally,
these would get auto-built out of grammar.jison. The correct regexes in
this case are simply putting optional negative signs in front of the
number regexes.